+++ controllers/ec.controller.js
@@ -3,7 +3,13 @@ const ec = {};
 // Requerimos pool de base de datos si es necesario
 const pool = require("../models/db");
 const { getUserDataByToken } = require("../middlewares/auth");
-const { Polly } = require("aws-sdk");
+const dayjs = require("dayjs");
+const isBetween = require("dayjs/plugin/isBetween");
+const customParseFormat = require("dayjs/plugin/customParseFormat");
+const { assign } = require("nodemailer/lib/shared");
+dayjs.extend(isBetween);
+dayjs.extend(customParseFormat);
+
 ec.main = (req, res) => {
   const usuario = getUserDataByToken(req.cookies.token);
   res.render("ec/formulario", usuario);
@@ -97,11 +103,11 @@ ec.addGrupo = async (req, res) => {
       [nombregrupo, id_carrera]
     );
     let modulos = pool.query(
-      "SELECT id , Nombre, fechaInicio, fechaFin FROM tb_ec_modulos WHERE isModel = 1 AND  idCarrera = ?",
+      "SELECT id , Nombre, fechaInicio, fechaFin , horas FROM tb_ec_modulos WHERE isModel = 1 AND  idCarrera = ?",
       id_carrera
     );
     let unidades = pool.query(
-      "SELECT id , Nombre, idModulo FROM tb_ec_unidades WHERE isModel = 1 AND  idCarrera = ?",
+      "SELECT id , Nombre, idModulo , horas FROM tb_ec_unidades WHERE isModel = 1 AND  idCarrera = ?",
       id_carrera
     );
     const query = await Promise.all([modulos, unidades]);
@@ -118,6 +124,7 @@ ec.addGrupo = async (req, res) => {
           obj["idUnidad"] = unidad.id;
           obj["Inicio"] = unidad.fechaInicio;
           obj["Fin"] = unidad.fechaFin;
+          obj["horas"] = unidad.horas;
           arrModelo.push(obj);
         }
       });
@@ -126,6 +133,7 @@ ec.addGrupo = async (req, res) => {
       obj["idModelo"] = element.id;
       obj["fechaInicio"] = element.fechaInicio;
       obj["fechaFin"] = element.fechaFin;
+      obj["horas"] = element.horas;
       datosModeloOrdenado[id] = obj;
     });
     const promesasUnidades = [];
@@ -136,14 +144,21 @@ ec.addGrupo = async (req, res) => {
       }
       let FechaInicio = `01/${element.fechaInicio}/${year}`;
       const idM = await pool.query(
-        "INSERT INTO tb_ec_modulos(Nombre,fechaInicio,fechaFin,Estado,isModel,idCarrera,idGrupo) VALUES(?,?,?,1,0,?,?)",
-        [element.modelo, FechaInicio, FechaFin, id_carrera, insertId]
+        "INSERT INTO tb_ec_modulos(Nombre,fechaInicio,fechaFin,horas,Estado,isModel,idCarrera,idGrupo) VALUES(?,?,?,?,1,0,?,?)",
+        [
+          element.modelo,
+          FechaInicio,
+          FechaFin,
+          element.horas,
+          id_carrera,
+          insertId,
+        ]
       );
       element.unidades.forEach((e) => {
         promesasUnidades.push(
           pool.query(
-            "INSERT INTO tb_ec_unidades(Nombre, isModel, idModulo, idCarrera, idGrupo)  VALUES(?,0,?,?,?)",
-            [e.UnidadName, idM.insertId, id_carrera, insertId]
+            "INSERT INTO tb_ec_unidades(Nombre, horas, isModel, idModulo, idCarrera, idGrupo)  VALUES(?,?,0,?,?,?)",
+            [e.UnidadName, e.horas, idM.insertId, id_carrera, insertId]
           )
         );
       });
@@ -169,7 +184,7 @@ ec.addModelo = async (req, res) => {
   }
 };
 ec.addUnidad = async (req, res) => {
-  const { Nombre, idModulo, idCarrera , horas } = req.body;
+  const { Nombre, idModulo, idCarrera, horas } = req.body;
   try {
     await pool.query(
       "INSERT INTO tb_ec_unidades(Nombre, horas , isModel, idModulo, idCarrera) VALUES (?, ?, 1 ,? ,?)",
@@ -202,13 +217,12 @@ ec.deleteUnidad = async (req, res) => {
   }
 };
 ec.editUnidad = async (req, res) => {
-  const { idUnidad, unidad  , horas} = req.body;
+  const { idUnidad, unidad, horas } = req.body;
   try {
-    await pool.query("UPDATE tb_ec_unidades SET Nombre=? , horas=? WHERE id=?  ", [
-      unidad,
-      horas,
-      idUnidad,
-    ]);
+    await pool.query(
+      "UPDATE tb_ec_unidades SET Nombre=? , horas=? WHERE id=?  ",
+      [unidad, horas, idUnidad]
+    );
     res.json({ status: true });
   } catch (error) {
     console.log(error);
@@ -216,7 +230,7 @@ ec.editUnidad = async (req, res) => {
   }
 };
 ec.editModulo = async (req, res) => {
-  const { idModulo, Nombre, Inicio, Fin , horas } = req.body;
+  const { idModulo, Nombre, Inicio, Fin, horas } = req.body;
   try {
     await pool.query(
       "UPDATE tb_ec_modulos SET Nombre=?  , fechaInicio=?, fechaFin = ?, horas = ?  WHERE id=?  ",
@@ -283,7 +297,7 @@ ec.administradorModelo = async (req, res) => {
           obj["UnidadID"] = unidad.id;
           obj["horas"] = unidad.horas;
           arrModelo.push(obj);
-          totalHorasUnidades = totalHorasUnidades+unidad.horas
+          totalHorasUnidades = totalHorasUnidades + unidad.horas;
         }
       });
       obj["modelo"] = element.Nombre;
@@ -301,11 +315,6 @@ ec.administradorModelo = async (req, res) => {
       carrera,
       carreraID: idCarrera,
     });
-    // res.json( {
-    //   datos: datosModeloOrdenado,
-    //   carrera,
-    //   carreraID: idCarrera,
-    // })
   } catch (error) {
     console.log(error);
     res.json({ status: false, error }).status(400);
@@ -320,11 +329,11 @@ ec.administradorCronogramaVigente = async (req, res) => {
       idGrupo
     );
     let modulos = pool.query(
-      "SELECT id , Nombre, fechaInicio, fechaFin FROM tb_ec_modulos WHERE isModel = 0 AND  idGrupo = ?",
+      "SELECT id , Nombre, fechaInicio, fechaFin, horas FROM tb_ec_modulos WHERE isModel = 0 AND  idGrupo = ?",
       idGrupo
     );
     let unidades = pool.query(
-      "SELECT tb_ec_unidades.id , tb_ec_unidades.Nombre, tb_ec_unidades.idModulo , tb_ec_unidades.fechaInicio, tb_ec_unidades.fechaFin, tb_ec_unidades.Estado , (SELECT Nombre FROM tb_usuarios WHERE id_usuario = tb_ec_unidades.id_usuario) AS Usuario FROM tb_ec_unidades WHERE isModel = 0 AND  idGrupo = ?",
+      "SELECT tb_ec_unidades.id , tb_ec_unidades.Nombre, tb_ec_unidades.idModulo , tb_ec_unidades.fechaInicio, tb_ec_unidades.fechaFin, tb_ec_unidades.Estado , (SELECT Nombre FROM tb_usuarios WHERE id_usuario = tb_ec_unidades.id_usuario) AS Usuario , tb_ec_unidades.horas FROM tb_ec_unidades WHERE isModel = 0 AND  idGrupo = ?",
       idGrupo
     );
     let profesores = pool.query(
@@ -341,6 +350,7 @@ ec.administradorCronogramaVigente = async (req, res) => {
     modulos.forEach((element, id) => {
       const arrModelo = [];
       const obj = {};
+      let horasTotalesUnidades = 0;
       unidades.forEach((unidad) => {
         if (unidad.idModulo == element.id) {
           let obj = {};
@@ -350,7 +360,9 @@ ec.administradorCronogramaVigente = async (req, res) => {
           obj["Fin"] = unidad.fechaFin;
           obj["Estado"] = unidad.Estado;
           obj["Usuario"] = unidad.Usuario;
+          obj["horas"] = unidad.horas;
           arrModelo.push(obj);
+          horasTotalesUnidades = horasTotalesUnidades + unidad.horas;
         }
       });
       obj["modelo"] = element.Nombre;
@@ -358,6 +370,8 @@ ec.administradorCronogramaVigente = async (req, res) => {
       obj["idModelo"] = element.id;
       obj["fechaInicio"] = element.fechaInicio;
       obj["fechaFin"] = element.fechaFin;
+      obj["horas"] = element.horas;
+      obj["horasUnidad"] = horasTotalesUnidades;
       datosModeloOrdenado[id] = obj;
     });
     res.render("ec/cronograma", {
@@ -374,11 +388,11 @@ ec.administradorCronogramaVigente = async (req, res) => {
   }
 };
 ec.editUnidadVigente = async (req, res) => {
-  const { idUnidad, unidad, inicio, fin, profesor } = req.body;
+  const { idUnidad, unidad, inicio, fin, profesor, horas } = req.body;
   try {
     await pool.query(
-      "UPDATE tb_ec_unidades SET Nombre=?, fechaInicio = ?, fechaFin =? , Estado = 1  , id_usuario=? WHERE id=?  ",
-      [unidad, inicio, fin, profesor, idUnidad]
+      "UPDATE tb_ec_unidades SET Nombre=?, fechaInicio = ?, fechaFin =? , Estado = 1  , id_usuario=? , horas = ? WHERE id=?  ",
+      [unidad, inicio, fin, profesor, horas, idUnidad]
     );
     res.json({ status: true });
   } catch (error) {
@@ -387,7 +401,7 @@ ec.editUnidadVigente = async (req, res) => {
   }
 };
 ec.addModeloVigente = async (req, res) => {
-  const { Nombre, Inicio, Fin, idGrupo } = req.body;
+  const { Nombre, Inicio, Fin, idGrupo, horas } = req.body;
   try {
     let idCarrera = await pool.query(
       "SELECT id_carrera FROM tb_ec_grupo WHERE id = ? ",
@@ -395,8 +409,8 @@ ec.addModeloVigente = async (req, res) => {
     );
     idCarrera = idCarrera[0].id_carrera;
     await pool.query(
-      "INSERT INTO tb_ec_modulos(Nombre, fechaInicio, fechaFin, Estado, isModel, idCarrera, idGrupo) VALUES (?, ? ,? , 1 , 0 , ? ,?)",
-      [Nombre, Inicio, Fin, idCarrera, idGrupo]
+      "INSERT INTO tb_ec_modulos(Nombre, fechaInicio, fechaFin , horas, Estado, isModel, idCarrera, idGrupo) VALUES (?, ? ,? , ?, 1 , 0 , ? ,?)",
+      [Nombre, Inicio, Fin, horas, idCarrera, idGrupo]
     );
     res.json({ status: true });
   } catch (error) {
@@ -405,7 +419,7 @@ ec.addModeloVigente = async (req, res) => {
   }
 };
 ec.addUnidadVigente = async (req, res) => {
-  const { Nombre, idModulo, profesor, inicio, fin, idGrupo } = req.body;
+  const { Nombre, idModulo, profesor, inicio, fin, idGrupo, horas } = req.body;
   try {
     let idCarrera = await pool.query(
       "SELECT id_carrera FROM tb_ec_grupo WHERE id = ? ",
@@ -413,8 +427,8 @@ ec.addUnidadVigente = async (req, res) => {
     );
     idCarrera = idCarrera[0].id_carrera;
     await pool.query(
-      "INSERT INTO tb_ec_unidades(Nombre, isModel, idModulo, idCarrera , fechaInicio, fechaFin, id_usuario , idGrupo, Estado) VALUES (?, 0 ,? ,?, ? , ?, ?,?, 1)",
-      [Nombre, idModulo, idCarrera, inicio, fin, profesor, idGrupo]
+      "INSERT INTO tb_ec_unidades(Nombre, isModel, idModulo, idCarrera , fechaInicio, fechaFin, horas ,id_usuario , idGrupo, Estado) VALUES (?, 0 ,? ,?, ? , ?, ?,? ,?, 1)",
+      [Nombre, idModulo, idCarrera, inicio, fin, horas, profesor, idGrupo]
     );
     res.json({ status: true });
   } catch (error) {
@@ -424,8 +438,201 @@ ec.addUnidadVigente = async (req, res) => {
 };
 
 ec.instructor = async (req, res) => {
-  const usuario = getUserDataByToken(req.cookies.token);
-  res.render("ec/instructores/carreras", usuario);
+  try {
+    const usuario = getUserDataByToken(req.cookies.token),
+      instructor = usuario.data.usuario,
+      promesas = [],
+      today = dayjs(),
+      format = "DD/MM/YYYY";
+    promesas.push(
+      pool.query(
+        "SELECT tb_ec_carrera.Nombre AS Carrera , tb_ec_unidades.idCarrera AS idCarrera, tb_ec_grupo.Nombre AS Grupo, tb_ec_grupo.id AS idGrupo    FROM `tb_ec_unidades` INNER JOIN tb_ec_carrera ON tb_ec_carrera.id = tb_ec_unidades.idCarrera INNER JOIN tb_ec_grupo ON tb_ec_grupo.id = tb_ec_unidades.idGrupo WHERE id_usuario = ? GROUP BY idGrupo",
+        [instructor]
+      )
+    );
+    promesas.push(
+      pool.query(
+        "SELECT tb_ec_unidades.id , tb_ec_modulos.Nombre AS modulo , tb_ec_unidades.Nombre , tb_ec_unidades.fechaInicio , tb_ec_unidades.fechaFin  , tb_ec_unidades.idCarrera , tb_ec_unidades.idGrupo FROM `tb_ec_unidades` INNER JOIN tb_ec_modulos ON tb_ec_modulos.id = tb_ec_unidades.idModulo WHERE tb_ec_unidades.id_usuario = ? AND tb_ec_unidades.isModel = 0",
+        [instructor]
+      )
+    );
+    const query = await Promise.all(promesas);
+    let unidades = query[1],
+      carreras = query[0],
+      unidadesValidas = [];
+    unidades.forEach((element) => {
+      const fechaInicio = dayjs(element.fechaInicio, format),
+        fechaFin = dayjs(element.fechaFin, format);
+      if (today.isBetween(fechaInicio, fechaFin.add(1, "day"), null, "[]"))
+        unidadesValidas.push(element);
+    });
+    const dataOrdenada = [];
+    carreras.forEach((carrera) => {
+      let unidades = [],
+        obj = {};
+      unidadesValidas.forEach((unidad) => {
+        if (unidad.idGrupo == carrera.idGrupo) {
+          unidades.push(unidad);
+        }
+        obj["unidades"] = unidades;
+        obj["Carrera"] = carrera.Carrera;
+        obj["idCarrera"] = carrera.idCarrera;
+        obj["grupo"] = carrera.Grupo;
+        obj["idGrupo"] = carrera.idGrupo;
+      });
+      dataOrdenada.push(obj);
+    });
+    res.render("ec/instructores/carreras", {
+      data: usuario.data,
+      dataOrdenada,
+    });
+  } catch (error) {
+    console.log(error);
+    res.status(400).json(error);
+  }
+};
+
+ec.contenidos = async (req, res) => {
+  const { unidad, grupo } = req.params,
+    usuario = getUserDataByToken(req.cookies.token);
+
+  try {
+    const evaluaciones = await pool.query(
+      "SELECT id, Descripcion, Tipo  FROM tb_ec_evaluaciones WHERE idUnidad = ?  ORDER BY Tipo",
+      [unidad]
+    );
+    res.render("ec/instructores/contenidos", {
+      data: usuario.data,
+      evaluaciones,
+      idUnidad: unidad,
+      grupo,
+    });
+  } catch (error) {
+    console.log(error);
+    res.status(400).json(error);
+  }
+};
+
+ec.newEva = async (req, res) => {
+  const { Nombre, Tipo, idUnidad } = req.body;
+  try {
+    if (!Nombre || !Tipo || !idUnidad) throw "PARAMS_NOT_COMPLETE";
+    const query = await pool.query(
+      "INSERT INTO tb_ec_evaluaciones(Descripcion, Tipo, idUnidad) VALUES (?,?, ?) ",
+      [Nombre, Tipo, idUnidad]
+    );
+    res.status(200).json({ status: true });
+  } catch (error) {
+    console.log(error);
+    res.status(400).json(error);
+  }
+};
+
+ec.deteleEva = async (req, res) => {
+  const { idEva } = req.body;
+  try {
+    if (!idEva) throw "PARAMS_NOT_COMPLETE";
+    await pool.query("DELETE FROM `tb_ec_evaluaciones` WHERE id  = ? ", [
+      idEva,
+    ]);
+    res.status(200).json({ status: true });
+  } catch (error) {
+    console.log(error);
+    res.status(400).json(error);
+  }
+};
+
+ec.notas = async (req, res) => {
+  const { evaluacion, grupo } = req.params,
+    usuario = getUserDataByToken(req.cookies.token),
+    promesas = [];
+  const { unidad } = req.query;
+
+  try {
+    promesas.push(
+      pool.query(
+        "SELECT id, Nombres, Apellidos FROM tb_ec_alumno WHERE id_grupo = ? ",
+        [grupo]
+      )
+    );
+    promesas.push(
+      pool.query(
+        "SELECT id, Nota, idAlumno, comentario FROM tb_ec_notas WHERE idEvaluacion = ? ",
+        [evaluacion]
+      )
+    );
+    const query = await Promise.all(promesas),
+      alumnos = query[0],
+      nota = query[1],
+      dataOrdenada = [];
+    alumnos.forEach((alumno, id) => {
+      let notaObtenida = {};
+      notaObtenida["Nota"] = 0;
+      notaObtenida["Comentario"] = "";
+      notaObtenida["Nombre"] = alumno.Nombres;
+      notaObtenida["Apellidos"] = alumno.Apellidos;
+      notaObtenida["idAlumno"] = alumno.id;
+      notaObtenida["isExist"] = false;
+      notaObtenida["idNota"] = false;
+      if (nota) {
+        let itnNota;
+        if (
+          nota.some((notaElement, it) => {
+            if (notaElement.idAlumno === alumno.id) {
+              itnNota = it;
+              return true;
+            }
+            return false;
+          })
+        ) {
+          notaObtenida["Nota"] = nota[itnNota].Nota;
+          notaObtenida["isExist"] = true;
+          notaObtenida["idNota"] = nota[itnNota].id;
+          notaObtenida["Comentario"] = nota[itnNota].comentario;
+        }
+      }
+      dataOrdenada.push(notaObtenida);
+    });
+    res.render("ec/instructores/notas", {
+      data: usuario.data,
+      dataOrdenada,
+      evaluacion,
+      grupo,
+      unidad,
+    });
+  } catch (error) {
+    console.log(error);
+    res.status(400).json(error);
+  }
+};
+
+ec.notasP = async (req, res) => {
+  const { notas } = req.body;
+  try {
+    let notasJson = JSON.parse(notas);
+    const promesas = [];
+    notasJson.forEach((nota) => {
+      if (nota.isExist) {
+        return promesas.push(
+          pool.query(
+            "UPDATE tb_ec_notas SET Nota = ?, comentario = ? WHERE idAlumno = ? AND id= ?",
+            [nota.Nota, nota.comentario, nota.alumnoId, nota.idNota]
+          )
+        );
+      }
+      promesas.push(
+        pool.query(
+          "INSERT INTO tb_ec_notas(Nota, comentario , idAlumno,idEvaluacion) VALUES (?,?,?,?) ",
+          [nota.Nota, nota.comentario, nota.alumnoId, nota.evaluacion]
+        )
+      );
+    });
+    await Promise.all(promesas);
+    res.status(200).json({ status: true });
+  } catch (error) {
+    console.log(error);
+    res.status(400).json(error);
+  }
 };
 
 module.exports = ec;